+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 - Fix missing image errors when deploying to Evolve
 - Fixed database migrations creating latest table state in the first stage instead of building incrementally
+- Cleaned up AcknowledgementLetterFromSource example
 
 ## [17.0.25] - 2026-06-15